docs(0004): replace account with from/to
Change the ledger plugin API to use from/to fields instead of the
ambiguous `account` field.

To illustrate how the previous API was confusing: In plugin-bells, we
actually populated the `account` field for incoming transfers with the
sending account, while we populated the `account` field for incoming
messages with the receiving (`to`) account.

Providing both sending and receiving accounts for both transfers and
messages allows us to more closely mirror the underlying API and reduce
the potential for confusion.