docs(0004,0009): Rename scale to currencyScale, deprecate currencySymbol, deprecate precision
For current ledgers, scale is always zero, so that's useless in ledgerInfo. However, currencyScale
is needed now. At the same time, deprecated currencySymbol since symbols, colors, and human
languages are a UI thing and not a concern of the ledger plugin. The plugin interface should only
deal with machine-readable data, not with human-computer interface.

See https://github.com/interledger/rfcs/issues/179 for a longer description of why precision is
being deprecated.