Remove regexes from table for MD compatibility
Since [GitHub changed their parser][1], [pipe characters in tables must now be escaped][2] even if they are part of inline code items. Since some other parsers don't work this way, the best way to solve this is to avoid using the regular expression in a table. Since the regexes for recognizing addresses are rather long anyway, taking them out of the table also reduces the chances they overflow off the side of the screen/column.

[1]: https://githubengineering.com/a-formal-spec-for-github-markdown/
[2]: https://github.github.com/gfm/#tables-extension-