docs(asn1): add Interledger RPC
slightly different take on:
https://github.com/interledger/rfcs/pull/251
and https://github.com/interledger/rfcs/pull/256
based on conversations with @michielbdejong, @sharafian, @justmoon

By separating the transfer contents from whether they are part of a
bilateral (e.g. trustline) or multilateral (e.g. centralized ledger)
relationship, the transfer protocol becomes much more general and thus
makes sense to include alongside the "core" ILP data formats.

InterledgerRpc wraps the InterledgerPacket types in an envelope that
adds a to, from (either or both of which can be set to empty strings if
they are implied, as is the case in a bilateral connection), a requestId
to correlate requests and responses, and sideProtocolData for
extensibility.

Any type of InterledgerRpc request can become a paid request by
including a Transfer or ConditionalTransfer in it (which themselves can
contain other InterledgerPackets -- InterledgerProtocolPayments or any
other type of data that you might want/need to pay for).