docs(asn1): add to/from to LLL request
The LLL is intended primarily for use in bilateral trustlines. However, we have added extensibility mechanisms that also allow it to be used as a base protocol for implementing Interledger over payment channels. There is only one feature that we need to make the same protocol also useful for multilateral ledgers: replacing the numeric  with  and  in the RPC wrapper. This is not strictly necessary but having a common base protocol that loads of people build off of (or implement directly in their ledgers) will save implementors a lot of time. It would also make debugging easier if the same tools could be used to inspect the messages of multiple ledger protocols. That is gained with minimal cost -- and actually switching from a 32-bit integer to two strings that can be empty saves 2 bytes when they're not being used -- so I think it's worth including the  and  fields.