docs(asn1): rename to CommonLedgerProtocol
Ledger-Ledger Lotocol is a bit too silly a name to keep. We should call this the Common Ledger Protocol, because it will be common in two senses of the word. First, we think it will be widely used. Second, we're already planning to use the sideProtocolData to extend this protocol into a lot of other ledger protocols. In that sense, this defines the common elements among a lot of different ledger protocols.

Also, when we had previously discussed the Common Ledger API in https://github.com/interledger/rfcs/pull/125, we didn't know enough about how the ledger protocol would be used to have an idea of what a common ledger interface would look like. I would argue that now we do, and this is our best guess at the common elements that will be used across many ledger protocols.