docs(clp): rm accountId
That feature is specifically for the admin plugin case, when you are authenticated as an admin user and acting on behalf of others. Since that is not an all the time case, that should go into the sideProtocolData. The reason for taking this out is also related to why we're not including https://github.com/interledger/rfcs/pull/261