docs(clp): rm TransferStateRequest
This is not needed in the core protocol. If you didn't hear a response to your request, you can just send it again and in all cases it should apply properly. The only case where you would want TransferStateRequest is where you remember the transfer ID but not what you wanted to do with it (i.e. either fulfill or reject). This isn't real. If this occurs, check for other glitches in the Matrix.