docs(clp): remove redundant response types
All of the response types are functionally equivalent, so this replaces them with three simple types that can be used for all requests. Note that we opted for having multiple types instead of using a type with a CHOICE in it because it means one more type to implement but one fewer bytes sent over the wire.

And then there were 8...