docs(3): add Amount Too Large error
connectors may want to limit the maximum size of transfers going through them. this adds an ILP error code to indicate when such a limit is reached. it also recommends including the amount that arrived and the maximum amount supported to help senders adjust the size of their payment appropriately