docs(1): ledger layer -> clearing + settlement layers
When we originally started working on Interledger, we thought of it
as a Real-Time Gross Settlement protocol. However, we have shfited to
thinking of it as primarily a protocol for clearing payments.
Nevertheless, we still refer to the layer below the Interledger layer as
the "Ledger Layer". This is confusing because it refers to both
settlement and clearing systems.

This commit splits what was formerly known as the Ledger Layer into the
Settlement Layer and Clearing Layer. This change should make it easier
to explain why we may assume _clearing_ is fast and inexpensive, even if
the underlying ledger or settlement system is slow or has high fees.