docs(asn1): less restrictive address format
We previously required ILP addresses to be at least length one, but for
the general type of `Address` it's actually better to allow zero-length,
so we can use it for prefixes in routing protocols etc.

The Interledger payment packets have been updated to match the previous
definition. The definition of the error packet has been intentionally
changed to allow empty addresses because the implementation allows and
is using this.