docs(25): remove chunked payments from PSKv2
Draft 1 attempted to have PSKv2 handle one-off payments, streaming payments, and chunked payments. However, @sharafian made good points that some of the features in the protocol and implementation of PSKv2 were only useful for chunked payments and not for streaming. When streaming payments, it makes more sense to have a completely stateless receiver that has a single callback to handle all incoming chunks, and applications will want more direct control over how they determine whether to accept or reject individual packets. Also, we wanted to [support application data](https://github.com/interledgerjs/ilp-protocol-psk2/issues/5) but PSKv2 did not go all the way to implementing segmentation and reassembly like TCP.

This change simplifies the protocol and includes only the features necessary for 1) generating the conditions and fulfillments from a shared secret 2) encrypting and authenticating request and response data and 3) providing the sender a way to indicate to the receiver the minimum amount they should get to accept a given packet.