Update 0009-simple-payment-setup-protocol.md
I made some inline comments in this document because I didn't know any other way. I don't expect you to approve and merge these as-is.

The main point of the comments is that SPSP is not an application layer protocol, and doesn't actually do what it is described as doing.